# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mistral.rs",
    "description": "Fast, flexible model inference.",
    "contact": {
      "name": "Eric Buehler"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.9.3"
  },
  "paths": {
...
    "/v1/completions": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "summary": "OpenAI-compatible completions endpoint handler.",
        "operationId": "completions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Completion JSON or server-sent event chunks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionChunkResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/embeddings": {
      "post": {
        "tags": [
          "Mistral.rs"
        ],
        "operationId": "embeddings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Embeddings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdapterSelection": {
        "oneOf": [
          {
            "type": "string",
            "description": "Resolve this alias when the request is admitted."
          },
          {
            "$ref": "#/components/schemas/AdapterGenerationSelection",
            "description": "Pin this exact resident generation."
          }
        ],
        "description": "A request-scoped LoRA adapter alias or exact immutable generation."
      },
...
      "CompletionChunkChoice": {
        "type": "object",
        "required": [
          "text",
          "index"
        ],
        "properties": {
          "finish_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "logprobs": {},
          "text": {
            "type": "string"
          }
        }
      },
      "CompletionChunkResponse": {
        "type": "object",
        "required": [
          "id",
          "choices",
          "created",
          "model",
          "system_fingerprint",
          "object"
        ],
        "properties": {
          "adapter_generation": {
            "type": [
              "string",
              "null"
            ]
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionChunkChoice"
            }
          },
          "created": {
            "type": "integer",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "system_fingerprint": {
            "type": "string"
          }
        }
      },
      "CompletionRequest": {
        "type": "object",
        "description": "Legacy OpenAI compatible text completion request",
        "required": [
          "prompt"
        ],
        "properties": {
          "adapter": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdapterSelection",
                "description": "Adapter alias or exact generation to activate for this request."
              }
            ]
          },
          "best_of": {
            "type": [
              "integer",
              "null"
            ],
            "example": 1,
            "minimum": 0
          },
          "dry_allowed_length": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Longest repeated sequence DRY leaves unpenalized.",
            "example": null,
            "minimum": 0
          },
          "dry_base": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Base for DRY's exponential penalty growth.",
            "example": null
          },
          "dry_multiplier": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "DRY repetition penalty multiplier; 0 disables DRY.",
            "example": null
          },
          "dry_sequence_breakers": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Sequences that reset DRY repetition matching.",
            "example": null
          },
          "echo": {
            "type": "boolean",
            "description": "Echo the prompt back alongside the completion.",
            "example": false
          },
          "frequency_penalty": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Penalize tokens by how often they have appeared so far; positive values reduce repetition.",
            "example": null
          },
          "grammar": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Grammar",
                "description": "Constrain output with a regex, JSON schema, Lark, or LLGuidance grammar."
              }
            ]
          },
          "ignore_eos": {
            "type": "boolean",
            "description": "Ignore model EOS tokens while still honoring explicit stops and the token limit.",
            "example": false
          },
          "logit_bias": {
            "type": [
              "object",
              "null"
            ],
            "description": "Bias added to the logits of these token IDs before sampling.",
            "additionalProperties": {
              "type": "number",
              "format": "float"
            },
            "propertyNames": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": null
          },
          "logprobs": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Include log probabilities of this many most likely tokens.",
            "example": null,
            "minimum": 0
          },
          "max_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Maximum number of tokens to generate.",
            "example": 16,
            "minimum": 0
          },
          "min_p": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Drop tokens below this fraction of the top token's probability.",
            "example": null
          },
          "model": {
            "type": "string",
            "description": "Model ID; \"default\" targets the only loaded model.",
            "example": "mistral"
          },
          "n": {
            "type": "integer",
            "description": "How many choices to generate.",
            "example": 1,
            "minimum": 0
          },
          "presence_penalty": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Penalize tokens that have already appeared; positive values push toward new topics.",
            "example": null
          },
          "prompt": {
            "type": "string",
            "example": "Say this is a test."
          },
          "repetition_penalty": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Multiplicative repetition penalty; 1.0 disables it.",
            "example": null
          },
          "seed": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Seed for deterministic request-scoped sampling.",
            "example": null,
            "minimum": 0
          },
          "stop": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StopTokens",
                "description": "Sequences where generation stops."
              }
            ]
          },
          "stream": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Stream the response as server-sent events."
          },
          "suffix": {
            "type": [
              "string",
              "null"
            ],
            "description": "Text appended after the completion.",
            "example": null
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Sampling temperature; higher values increase randomness.",
            "example": 0.7
          },
          "tool_choice": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ToolChoice",
                "description": "Controls which (if any) tool the model must call."
              }
            ]
          },
          "tools": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Tool"
            },
            "description": "Tools the model may call.",
            "example": null
          },
          "top_k": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Sample only from the k most likely tokens.",
            "example": null,
            "minimum": 0
          },
          "top_p": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Nucleus sampling: only tokens within the top cumulative probability mass are considered.",
            "example": null
          },
          "truncate_sequence": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Truncate inputs that exceed the model's context length instead of erroring.",
            "example": null
          },
          "user": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CompletionResponse": {
        "type": "object",
        "required": [
          "id",
          "choices",
          "created",
          "model",
          "system_fingerprint",
          "object",
          "usage"
        ],
        "properties": {
          "adapter_generation": {
            "type": [
              "string",
              "null"
            ]
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletionResponseChoice"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "system_fingerprint": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/CompletionUsageResponse"
          }
        }
      },
      "CompletionResponseChoice": {
        "type": "object",
        "required": [
          "finish_reason",
          "index",
          "text"
        ],
        "properties": {
          "finish_reason": {
            "type": "string"
          },
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "logprobs": {},
          "text": {
            "type": "string"
          }
        }
      },
...
      "EmbeddingEncodingFormat": {
        "type": "string",
        "enum": [
          "float",
          "base64"
        ]
      },
...
      "EmbeddingRequest": {
        "type": "object",
        "required": [
          "input"
        ],
        "properties": {
          "dimensions": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Truncate embeddings to this dimensionality, if the model supports it.",
            "example": null,
            "minimum": 0
          },
          "encoding_format": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EmbeddingEncodingFormat",
                "description": "Return embeddings as float arrays or base64-encoded strings."
              }
            ]
          },
          "input": {
            "$ref": "#/components/schemas/EmbeddingInput",
            "description": "Text or token inputs to embed."
          },
          "model": {
            "type": "string",
            "description": "Model ID; \"default\" targets the only loaded model.",
            "example": "default"
          },
          "truncate_sequence": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Truncate inputs that exceed the model's context length instead of erroring.",
            "example": null
          },
          "user": {
            "type": [
              "string",
              "null"
            ],
            "example": null
          }
        }
      },
      "EmbeddingResponse": {
        "type": "object",
        "required": [
          "object",
          "data",
          "model",
          "usage"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmbeddingData"
            }
          },
          "model": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/EmbeddingUsage"
          }
        }
      },
...
      "StopTokens": {
        "oneOf": [
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Multiple possible stop sequences"
          },
          {
            "type": "string",
            "description": "Single stop sequence"
          }
        ],
        "description": "Stop token configuration for generation\n\nDefines when the model should stop generating text, either with a single\nstop token or multiple possible stop sequences."
      },
...
      "Tool": {
        "type": "object",
        "description": "Tool definition",
        "required": [
          "type",
          "function"
        ],
        "properties": {
          "function": {
            "$ref": "#/components/schemas/Function"
          },
          "type": {
            "$ref": "#/components/schemas/ToolType"
          }
        }
      },
...
      "ToolChoice": {
        "oneOf": [
          {
            "type": "string",
            "description": "Disallow selection of tools.",
            "enum": [
              "none"
            ]
          },
          {
            "type": "string",
            "description": "Allow automatic selection of any given tool, or none.",
            "enum": [
              "auto"
            ]
          },
          {
            "type": "string",
            "description": "Require at least one tool call.",
            "enum": [
              "required"
            ]
          },
          {
            "$ref": "#/components/schemas/AllowedToolsToolChoice"
          },
          {
            "$ref": "#/components/schemas/BuiltinToolChoice"
          },
          {
            "$ref": "#/components/schemas/Tool"
          },
          {
            "$ref": "#/components/schemas/NamedFunctionToolChoice"
          }
        ]
      },
...
    }
  },
...
}